Moved Founders tests to their own file.
